# openapi.yaml
openapi: 3.0.3
info:
  title: Poolantir Simulation Controller API
  version: 1.0.0
paths:
  /api/v1/state:
    get:
      summary: Get state snapshot
      responses:
        "200":
          description: Runtime state
  /api/v1/config:
    get:
      summary: Get config
      responses:
        "200":
          description: Config object
    put:
      summary: Update config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stalls:
                  type: integer
                urinals:
                  type: integer
                pee_duration_sec:
                  type: integer
                poo_duration_sec:
                  type: integer
      responses:
        "200":
          description: Updated config
  /api/v1/demand/delta:
    post:
      summary: Apply queue delta
      parameters:
        - in: header
          name: Idempotency-Key
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pee_delta:
                  type: integer
                poo_delta:
                  type: integer
      responses:
        "200":
          description: Updated queue
  /api/v1/incidents:
    get:
      summary: List incidents
      responses:
        "200":
          description: List incidents
    post:
      summary: Open an incident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Created incident
  /api/v1/incidents/{incident_id}:
    patch:
      summary: Resolve incident
      parameters:
        - in: path
          name: incident_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated incident
  /api/v1/nodes/register:
    post:
      summary: Register node
      responses:
        "201":
          description: Registered node
  /api/v1/nodes/{node_id}/heartbeat:
    post:
      summary: Node heartbeat
      parameters:
        - in: path
          name: node_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated node
  /api/v1/nodes/{node_id}/ack-start:
    post:
      summary: Ack assignment start
      responses:
        "200":
          description: Acked
  /api/v1/nodes/{node_id}/ack-done:
    post:
      summary: Ack assignment done
      responses:
        "200":
          description: Acked
  /api/v1/nodes/{node_id}/ack-fail:
    post:
      summary: Ack assignment failed
      responses:
        "200":
          description: Acked
  /api/v1/stream:
    get:
      summary: SSE stream endpoint
      responses:
        "200":
          description: text/event-stream payload
  /api/v1/simulation/usage:
    post:
      summary: Ingest simulation restroom usage events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/RestroomUsageEvent"
                - type: object
                  properties:
                    source:
                      type: string
                      enum: [simulation, dummy]
                    events:
                      type: array
                      minItems: 1
                      items:
                        $ref: "#/components/schemas/RestroomUsageEvent"
      responses:
        "201":
          description: Usage events ingested
        "400":
          description: Invalid payload
  /api/v1/simulation/dummy/generate:
    post:
      summary: Generate synthetic anomaly and baseline usage records
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                count:
                  type: integer
                  minimum: 1
                  maximum: 2000
                  default: 25
                restroom:
                  type: string
                  description: Optional restroom alias or canonical restroom id
                seed:
                  type: integer
                  description: Optional deterministic random seed
      responses:
        "201":
          description: Synthetic usage events generated
        "400":
          description: Invalid payload
components:
  schemas:
    RestroomUsageEvent:
      type: object
      required: [restroom, node_id, toilet_type, duration_s]
      properties:
        restroom:
          type: string
          description: Alias or canonical id (seamen/seamans/maclean or canonical restroom id)
        restroom_id:
          type: string
          description: Canonical restroom id alternative to restroom
        node_id:
          type: integer
        toilet_type:
          type: string
          enum: [stall, urinal]
        duration_s:
          type: number
          format: float
        scenario:
          type: string
        is_anomaly:
          type: boolean